pyproject.toml
[project]
name = "pdf-o-matic"
version = "0.1.0"
description = "ILovePDF-kind app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "reflex==0.8.17a1",
    "pypdf==6.1.3",
    "pymupdf==1.26.5",
    "pillow==12.0.0",
    "ruff>=0.14.2",
]
